création des pages FicheUser, Home, et du component layout avec fichier MainHeader
