+++ frontend/src/App.jsx
@@ -1,7 +1,8 @@
 import React, { useContext } from "react";
-import AuthForm from "@components/Auth/AuthForm";
-import Test from "@components/test";
 import AuthContext from "./context/auth_context";
+import Home from "@pages/Home";
+import FicheUser from "@pages/FicheUser";
+import MainHeader from "@components/layout/MainHeader";
 // import { BrowserRouter as Router, Route, Routes} from "react-router-dom";
 
 function App() {
@@ -9,8 +10,11 @@ function App() {
   const isLoggedIn = authCtx.isLoggedIn;
   return (
     <div className="App">
-      {!isLoggedIn && <AuthForm />}
-      <Test />
+      <Home />
+      <FicheUser />
+      <MainHeader />
+      {/* {!isLoggedIn && <AuthForm />}
+      <Test /> */}
       {/* <Router>
         <Routes>
           <Route path="connexion" element={< />} />
