+++ frontend/src/components/Auth/AuthForm.jsx
@@ -82,7 +82,7 @@ return false;
           email: "",
           password: "",
         });
-        setSuccesMessage(` Bienvenue ${res.data.result[0].firstname} vous êtes bien connecté`);
+        setSuccesMessage(` Bienvenue ${res.data.result[0].lastname} vous êtes bien connecté`);
         setShowSuccesModal(true);
 
         
@@ -116,7 +116,7 @@ return false;
       email: "",
       password: "",
      })
-      setSuccesMessage(` Bienvenue ${res.data.result[0].firstname} vous êtes bien enregisté`);
+      setSuccesMessage(` Bienvenue ${res.data.result[0].lastname} vous êtes bien enregisté`);
       setShowSuccesModal(true);
       
     } catch (err) {
