+++ frontend/src/components/layout/MainHeader.jsx
@@ -1,0 +1,14 @@
+import React from 'react'
+
+const MainHeader = () => {
+  return (
+    <header>
+      <nav>
+        
+      </nav>
+    </header>
+    
+  )
+}
+
+export default MainHeader;
