+++ frontend/src/pages/Home.jsx
@@ -1,0 +1,12 @@
+import AuthForm from '@components/Auth/AuthForm'
+import React from 'react'
+
+const Home = () => {
+  return (
+   <section>
+    <AuthForm />
+   </section>
+  )
+}
+
+export default Home;
