+++ frontend/src/styles/_setting.scss
@@ -20,11 +20,8 @@ $font-3: 'Roboto', sans-serif;
 }
 
 body{
-  width: 100vw;
-  height: 100vh;
-  display: flex;
-  justify-content: center;
-  align-items: center;
+ display:flex;
+ justify-content: center;
   font-family: $font-3;
   background-color:rgb(15, 9, 48)
