+++ frontend/src/styles/_test.scss
@@ -1,6 +1,10 @@
 .test{
-  width: 50px;
-  font-size: 1rem;
-  margin-top:10rem;
+//   // width: 50px;
+  font-size: 1.5rem;
+//   margin-top:5rem;
   color :white;
+}
+
+.token{
+  white-space: pre-wrap;
 }
